+++ ChangeLog.md
@@ -1,4 +1,8 @@
 # Version history for fedora-composes
 
+## 0.2 (2022-xx-yy)
+- output compose urls
+- spread status over several lines
+
 ## 0.1 (2022-08-11)
 - initial release with list and status commands and filtering